Refactor audio module structure; reorganize imports and create new __init__.py files for core, models, and runnables
