+++ scinoephile/audio/core/__init__.py
@@ -1,0 +1,15 @@
+#  Copyright 2017-2025 Karl T Debiec. All rights reserved. This software may be modified
+#  and distributed under the terms of the BSD license. See the LICENSE file for details.
+"""Core code related to audio."""
+
+from __future__ import annotations
+
+from scinoephile.audio.core.audio_block import AudioBlock
+from scinoephile.audio.core.audio_series import AudioSeries
+from scinoephile.audio.core.audio_subtitle import AudioSubtitle
+
+__all__ = [
+    "AudioBlock",
+    "AudioSeries",
+    "AudioSubtitle",
+]
